Disable installation during build phase
This is to save the ~5GB on image size from importing things like PyTorch.
The downside is that means these will now have to be downloaded when the image
is ran. Hopefully this can be offset by saving the heavy items to an anonymous
volume. Without that anonymous volume however, this will mean every time the
image is rebooted, it will redownload the 5GB worth of dependencies. Fun stuff.